requirements.txt
Django==3.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
whitenoise==5.2.0
djangorestframework
django-cors-headers
PyPDF2
django-filter
